Update image of extracted otolith
The notebook now uses connected components to extract the four largest
items in the cropped dataset.
A simple zero-trimming crops the data even further, making it easily
visualizable.
Add a screenshot of the visualization to the manuscript.
The visualization needs to be updated with proper scale bars and more
pizzazz.